+++ objects/wrapper/src/packer_wrapper/pack_thread_to_json.c
@@ -1,0 +1,38 @@
+/*
+** EPITECH PROJECT, 2022
+** my_teams
+** File description:
+** pack_thread_to_json
+*/
+
+#include "packer_json.h"
+
+static void thread_to_json(thread_t *thread, FILE *file)
+{
+    char time_display[20];
+
+    fprintf(file, "{\"UUID\":\"%s\",", thread->uuid->uuid.repr);
+    fprintf(file, "\"Parent\":\"%s\",", thread->channel->uuid.repr);
+    fprintf(file, "\"Author\":\"%s\",", thread->author->uuid.repr);
+    fprintf(file, "\"Name\":\"%s\",", thread->name);
+    fprintf(file, "\"Body\":\"%s\",", thread->body);
+    strftime(time_display, 20, "%Y-%m-%d %H:%M:%S", localtime(&thread->date));
+    fprintf(file, "\"Time\":\"%s\",", time_display);
+    fprintf(file, "\"Comments\":[");
+    for (int i = 0; i < thread->comment_n; i++) {
+        fprintf(file, "{\"UUID\":\"%s\"}", thread->comments[i]->uuid.repr);
+        if (i + 1 < thread->comment_n)
+            fprintf(file, ",");
+    }
+    fprintf(file, "]}");
+}
+
+void pack_wrapper_to_json_thread(object_wrapper_t *wrapper, FILE *file)
+{
+    fprintf(file, "],\"Thread number\":%i,\"Threads\":[", wrapper->thread_n);
+    for (int i = 0; i < wrapper->thread_n; i++) {
+        thread_to_json(wrapper->threads[i], file);
+        if (i + 1 < wrapper->thread_n)
+            fprintf(file, ",");
+    }
+}
